Fix a few bugs
When a user were to hit the equal sign and
pressed a number when they got their answer,
it would continue adding the pressed number to
the end of the answer.

A minor fix for the css and html where if a user
gets a negative number, the negative sign will
show up behind the number.