Fixed bug where user can add to the answer again
It was possible for the user to add to the answer
but was fixed previously.

The next bug was when the user hit a number and
any operator following that number. If the user
kept alternating between number and operator it
would break the calculator.

Lastly would be that I implemented a way to make
decimals by adding the period. Now the user can
make any decimal numbers.